Healthbar fixed
Healthbar fixed

+++ Assets/PlayerHealthbar.cs
@@ -6,14 +6,15 @@ using UnityEngine.UI;
 public class PlayerHealthbar : MonoBehaviour
 {
   public Slider slider;
-  public void SetMaxHealth(int health)
+
+  public void SetMaxHealth(float health)
   {
-slider.maxValue = health;
-slider.value = health;
+    slider.maxValue = health;
+    slider.value = health;
   }
 
-  public void SetHealth (int health)
+  public void SetHealth (float health)
   {
-slider.value = health;
+    slider.value = health;
   }
 }
